test: add dedicated test file for product filter stock availability
